Move purchase form functionality to a separate module
